+++ src/main/java/com/project/coalba/domain/substituteReq/service/BossSubstituteReqService.java
@@ -1,11 +1,5 @@
 package com.project.coalba.domain.substituteReq.service;
 
-import com.project.coalba.domain.auth.entity.User;
-import com.project.coalba.domain.externalCalendar.dto.CalendarEventDto;
-import com.project.coalba.domain.externalCalendar.dto.CalendarPersonalDto;
-import com.project.coalba.domain.externalCalendar.service.ExternalCalendarService;
-import com.project.coalba.domain.notification.FirebaseCloudMessageService;
-import com.project.coalba.domain.profile.entity.Staff;
 import com.project.coalba.domain.schedule.entity.Schedule;
 import com.project.coalba.domain.substituteReq.dto.response.BothSubstituteReqResponse;
 import com.project.coalba.domain.substituteReq.dto.response.YearMonth;
@@ -26,7 +20,6 @@ import static java.util.stream.Collectors.groupingBy;
 @RequiredArgsConstructor
 @Service
 public class BossSubstituteReqService {
-    private final FirebaseCloudMessageService firebaseCloudMessageService;
     private final SubstituteReqRepository substituteReqRepository;
     private final ProfileUtil profileUtil;
 
@@ -60,17 +53,7 @@ public class BossSubstituteReqService {
         Schedule schedule = substituteReq.getSchedule();
         schedule.changeStaff(substituteReq.getReceiver());
 
-        sendApprovalNotice(substituteReq);
         return substituteReq;
-    }
-
-
-    private void sendApprovalNotice(SubstituteReq substituteReq) {
-        String senderTargetToken = substituteReq.getSender().getDeviceToken();
-        String receiverTargetToken = substituteReq.getReceiver().getDeviceToken();
-
-        firebaseCloudMessageService.sendMessageTo(senderTargetToken, "대타 승인", "스케줄에 해당 근무가 삭제되었습니다.");
-        firebaseCloudMessageService.sendMessageTo(receiverTargetToken, "대타 승인", "스케줄에 해당 근무가 추가되었습니다.");
     }
 
     @Transactional
